mute_site/requirements.txt
Django==3.1.6
djangorestframework
pillow